WOM-392: Update int arg types and try nodeid before read_browse_name()
Resolves issue with recursive scanning of PLC's
"Pointing.StaticPmParameters" tree not behaving the same as it does with
the CETC sim. The BrowseName of the PLC's nodes do not have [0]..[7] at
the end like the node ID, but SCU does not need to use the
BrowseName, as the name can be parsed from nodeid.Identifier
already. Doing this removes one asyncio call for each node during tree
scanning.

GUI also automatically releases command authority when disconnecting
from server.

+++ src/disq/model.py
@@ -6,6 +6,7 @@ from enum import Enum
 from pathlib import Path
 from queue import Empty, Queue
 
+from asyncua.ua import UInt16
 from PyQt6.QtCore import QObject, QThread, pyqtSignal
 
 from disq.logger import Logger
@@ -289,11 +290,13 @@ class Model(QObject):
                 band = self._scu.convert_enum_to_int("BandType", args[3])
                 static = args[0]
                 tilt = self._scu.convert_enum_to_int("TiltOnType", args[1])
-                if tilt is None:  # TODO: Remove once PLC is fixed
+                if tilt is None:  # TODO: Remove 'if' block once PLC is fixed
                     logger.warning(
                         "OPC-UA server has no 'TiltOnType' enum. Attempting a guess."
                     )
-                    tilt = {"Off": 0, "TiltmeterOne": 1, "TiltmeterTwo": 2}[args[1]]
+                    tilt = UInt16(
+                        {"Off": 0, "TiltmeterOne": 1, "TiltmeterTwo": 2}[args[1]]
+                    )
                 temperature = args[2]
                 result = _log_and_call(command, static, tilt, temperature, band)
             case Command.TAKE_AUTH:
